# Clean up array manipulation