--- a/app/build.gradle
+++ b/app/build.gradle
@@ -4,7 +4,7 @@ android {
     compileSdkVersion 29
     defaultConfig {
         applicationId "com.getbouncer.testcardscan"
-        minSdkVersion 19
+        minSdkVersion 21
         targetSdkVersion 29
         versionCode 4015
         versionName "1.0.5139"