--- a/cardscan-base/src/main/java/com/getbouncer/cardscan/base/ObjectDetect.java
+++ b/cardscan-base/src/main/java/com/getbouncer/cardscan/base/ObjectDetect.java
@@ -48,7 +48,7 @@ public class ObjectDetect {
     }
 
     private void ssdOutputToPredictions(@NonNull Bitmap image) {
-        if (ssdDetect == null) {
+        if (ssdDetect == null || priors == null) {
             return;
         }