I spent some time going through the array utils and cleaning them up. I timed the execution of each method, each is either the same or slightly faster than the original.

This allows us to address individual values in the matrices:
```
float[] location = sizeAndCenter(1.0, 2.0, 3.0, 4.0)
location.centerX() == 1.0
location.centerY() == 2.0
location.width() == 3.0
location.height() == 4.0

location.setCenterX(5.0)
...

location.toRectForm().left() == -0.5
location.toRectForm().right() == 2.5
...
```

All this makes understanding the matrix math much easier.

# TestAndroidOcr results

## This branch
|  | pixel 2 xl | nexus 5x |
| -- | -------- | ------------- |
| avg pred (ms) | 132 | 242 |
| frames | 640 | 640 |
| correct | 301 | 299 |
| incorrect | 0 | 0 |

## Master branch (1.0.5139)
|  | pixel 2 xl | nexus 5x |
| -- | -------- | ------------- |
| avg pred (ms) | 126 | 294 |
| frames | 640 | 640 |
| correct | 301 | 299 |
| incorrect | 0 | 0 |